Handles ban and fav (?!) case

+++ source/feedback.py
@@ -7,11 +7,9 @@ import twelec_globals
 def feedback(request):
 
     session_id=int(request.form['session_id'])
-    for key in request.form.keys():    
-        print(key)
 
     # Browse all kept tweets and check wether
-    # some are marked as to ban
+    # some are marked as to ban or faved
     with lite.connect("twitter.db") as con:
 
         cur=con.cursor()
@@ -21,14 +19,27 @@ def feedback(request):
         cur.execute("SELECT TwId FROM KeptTweets WHERE Session=?",(session_id,))
 
         row=cur.fetchone()
+        faved_tweets=[]
         while row != None:
+            # If the tweet is banned, remove it from the kept table and tag it as banned in the fetched table
             try:
-                if request.form['ban_'+str(row[0])]:                
+                if request.form['ban_'+str(row[0])]=="on" and 'fav_'+str(row[0]) not in (request.form).keys():                
                     cur_update.execute("DELETE FROM KeptTweets WHERE TwId=?",(row[0],))
                     cur_update.execute("UPDATE FetchedTweets SET State=? WHERE TwID=?",(twelec_globals.tweet_states['banned'],row[0]))
             except KeyError:
                 pass
+
+            # If the tweet is faved, add it to the faved list
+            try:
+                if request.form['fav_'+str(row[0])]=="on" and 'ban_'+str(row[0]) not in (request.form).keys():
+                    cur_update.execute("UPDATE FetchedTweets SET State=? WHERE TwID=?",(twelec_globals.tweet_states['faved'],row[0]))               
+                    faved_tweets.append(row[0])
+            except KeyError:
+                pass
+
             
             row=cur.fetchone()
-            
+
+        # Process faved tweets
+                        
     return(displayTweets.displayToStr(session_id))
